+++ src/Transformer/TextContainsTransformer.php
@@ -30,11 +30,11 @@ class TextContainsTransformer extends AbstractTransformer
     protected function transform() : void
     {
         $keyLocator = array_keys($this->step['textContains'])[1];
-        $errorMessage = sprintf('%s[%s] text does not contains %s', $keyLocator, $this->step['textContains'][$keyLocator], $this->step['value']);
+        $errorMessage = sprintf('%s[%s] text does not contains %s', $keyLocator, $this->step['textContains'][$keyLocator], $this->step['textContains']['value']);
         $this->driver->wait(Configuration::get('wait.for'), Configuration::get('wait.every'))
                      ->until(WebDriverExpectedCondition::elementTextIs(
-                         WebLocator::get($keyLocator, array_values($this->step['textContains'])[$keyLocator]),
-                         $this->step['value']
+                         WebLocator::get($keyLocator, $this->step['textContains'][$keyLocator]),
+                         $this->step['textContains']['value']
                      ), $errorMessage);
     }