+++ src/Transformer/TextIsTransformer.php
@@ -30,11 +30,11 @@ class TextIsTransformer extends AbstractTransformer
     protected function transform() : void
     {
         $keyLocator = array_keys($this->step['textIs'])[1];
-        $errorMessage = sprintf('%s[%s] text is not %s', $keyLocator, $this->step['textIs'][$keyLocator], $this->step['value']);
+        $errorMessage = sprintf('%s[%s] text is not %s', $keyLocator, $this->step['textIs'][$keyLocator], $this->step['textIs']['value']);
         $this->driver->wait(Configuration::get('wait.for'), Configuration::get('wait.every'))
                      ->until(WebDriverExpectedCondition::elementTextIs(
-                         WebLocator::get($keyLocator, array_values($this->step['textIs'])[$keyLocator]),
-                         $this->step['value']
+                         WebLocator::get($keyLocator, $this->step['textIs'][$keyLocator]),
+                         $this->step['textIs']['value']
                      ), $errorMessage);
     }