+++ src/Transformer/TextMatchesTransformer.php
@@ -30,11 +30,11 @@ class TextMatchesTransformer extends AbstractTransformer
     protected function transform() : void
     {
         $keyLocator = array_keys($this->step['textMatches'])[1];
-        $errorMessage = sprintf('%s[%s] text does not match %s', $keyLocator, $this->step['textMatches'][$keyLocator], $this->step['value']);
+        $errorMessage = sprintf('%s[%s] text does not match %s', $keyLocator, $this->step['textMatches'][$keyLocator], $this->step['textMatches']['regexp']);
         $this->driver->wait(Configuration::get('wait.for'), Configuration::get('wait.every'))
                      ->until(WebDriverExpectedCondition::elementTextMatches(
-                         WebLocator::get($keyLocator, array_values($this->step['textMatches'])[$keyLocator]),
-                         $this->step['regexp']
+                         WebLocator::get($keyLocator, $this->step['textMatches'][$keyLocator]),
+                         $this->step['textMatches']['regexp']
                      ), $errorMessage);
     }