+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Scenario` can now reset or not the scope. When using `use`, the Scenario variable scope is not reset.
 - When AutoMate runs, the World variable scope is reset.
 - `Use` command now use the good scenario.
+- `textIs`, `textContains`, `textMatches` work now
+- Fixed a bug that throw an error when the `GetTransformer` set the World variable
 
 ## [0.4.1] - 28/02/2021